Updating comments on file
Still cleaning up comments and code; spacing, line breaks, tabs, etc.

Create README.md

Final update of README.md (I hope), adding MIT License

still fine tuning readme
