+++ README.md
@@ -1,0 +1,45 @@
+# Alimentum - Coding Dojo iOS Project
+
+######Credits
+
+ >Developed by Nitish Dayal, Joseph Park <br />
+ >Coding Dojo -  Graduating Cohort March 2016
+
+###Core Functionality:
+  Request User location and generate list of 10 closest open food-related businesses in the area, 
+  sortable by "Delivery" or "All".
+  User can call businesses listed (if valid number is provided) directly from application, and will be returned to application on call end.
+
+  
+### Resources Used:
+
+######Development Environment: 
+  - Swift 2.2
+  - Xcode 7.3
+  - Yelp API
+  - OAuthSwift
+  - iOS Core Frameworks: Core Data
+  - Github and Git
+  
+######Research & Reference Material: 
+  - The Swift Programming Language - Swift 2.2 Edition
+  - Apple Developer Library
+  - Stack Overflow
+  - Github Search
+  - Yelp API Documentation
+  
+###About:
+  This application was designed and developed between Tuesday, 3/29/2016, and Friday, 4/01/2016, by Nitish Dayal & Joseph Park. The primary intent behind developing this application was to gain a better understanding of what was possible with the knowledge we obtained at Coding Dojo. Througout the development process, we utilized Git & Github for version control, Asana for task delegation, and many resources (both internal and external) to put together the necessary bits and pieces. While the current version is by no means complete, continued work on this project is being debated. The primary intent of this project was met; we set out to develop something that we could present during Project Week, and that was a success. Both Joseph & I have other project ideas we are interested in getting started on. There is also the matter of the Yelp API ToC, which essentially require any use of their API to be followed with a barrage of links/ads/logos.
+  
+However we did not get a chance to finish the application the way __we wanted to__, which, frankly, is far more important to me. We want more usability: what if we allowed the user to select between different category options (ie; food, parks, bars, events, etc.), or choose how many results they wanted returned, or save generated lists under location-based categories (ie; home-food-list, school-parks-list, work-bars-list, etc.) to allow for some offline functionality? We wanted to integrate Social Networking/Sharing services. We wanted _pictures_. So as it stands, this project is best defined as "on hiatus", though there is a strong possibility that we will return and add at least one or two final touches within the coming weeks. When that time comes, here is a brief list of ideas to implement:
+
+######TODO:
+    - [ ] Use CoreData to save offline lists
+    - [ ] Find a way to do ^ without breaking Yelp ToC
+    - [ ] Add drop-down list of basic category options
+    - [ ] Change Main View Controller title label to update value based on current category 
+    - [ ] Add some images. Somewhere. Anywhere. Not really, it should be tasteful, but yes we need images.
+    - [ ] Decide when to finally walk away from this.
+
+###License:
+  Released under MIT License
