+++ alimentum/alimentum/BusinessListingTableViewCell.swift
@@ -1,13 +1,17 @@
 //
-//  BusinessListingTableViewCell.swift
+//  AppDelegate.swift
 //  alimentum
 //
-//  Created by Nitish Dayal on 3/29/16.
-//  Copyright © 2016 Joseph Park. All rights reserved.
+//  Created by Nitish Dayal, Joseph Park
+//  Copyright © 2016 Nitish Dayal, Joseph Park. All rights reserved.
+//
+//  Libraries used in project include: UIKit, CoreLocation, 0AuthSwift
 //
 
 import UIKit
 
+
+//MARK: - Define custom cell in MainViewController
 class BusinessListingTableViewCell: UITableViewCell {
 
     @IBOutlet weak var businessName: UILabel!
@@ -15,7 +19,5 @@ class BusinessListingTableViewCell: UITableViewCell {
     @IBOutlet weak var address: UILabel!
     @IBOutlet weak var rating: UILabel!
     @IBOutlet weak var callPhoneNumber: UIButton!
-    
-
 
 }
