+++ alimentum/alimentum/YelpClient.swift
@@ -1,24 +1,31 @@
 //
-//  YelpClient.swift
+//  AppDelegate.swift
 //  alimentum
 //
-//  Created by Nitish Dayal on 3/30/16.
-//  Copyright © 2016 Joseph Park. All rights reserved.
+//  Created by Nitish Dayal, Joseph Park
+//  Copyright © 2016 Nitish Dayal, Joseph Park. All rights reserved.
+//
+//  Libraries used in project include: UIKit, CoreLocation, 0AuthSwift
 //
 
 import UIKit
 import OAuthSwift
 
+//MARK: - YelpAPIConsole struct, to hold key & secret values for API call
 struct YelpAPIConsole {
+    
     var consumerKey = Credentials().c_key
     var consumerSecret = Credentials().c_secret
     var accessToken = Credentials().a_Token
     var accessTokenSecret = Credentials().a_TokenSecret
+    
 }
 
+
+//MARK: - Make custom class YelpAPIClient that conforms to class NSObject
 class YelpAPIClient: NSObject {
     
-    let APIBaseUrl = "https://api.yelp.com/v2"
+    let yelpSearchUrl = "https://api.yelp.com/v2/search"
     let clientOAuth: OAuthSwiftClient?
     let apiConsoleInfo: YelpAPIConsole
     
@@ -28,9 +35,8 @@ class YelpAPIClient: NSObject {
         super.init()
     }
 
-    
+    /* Add GET request functionality with Yelp! API authorized search URL */
     func searchPlacesWithParameters(searchParameters: Dictionary<String, String>, successSearch: (data: NSData, response: NSHTTPURLResponse) -> Void, failureSearch: (error: NSError) -> Void) {
-        let searchUrl = APIBaseUrl + "/search"
-        clientOAuth!.get(searchUrl, parameters: searchParameters, success: successSearch, failure: failureSearch)
+        clientOAuth!.get(yelpSearchUrl, parameters: searchParameters, success: successSearch, failure: failureSearch)
     }
 }
